feat(handler): 统一集群ID参数类型为Long，提升数据一致性
- 将多个处理类中的集群ID参数类型从Integer更改为Long，以确保与数据模型的一致性。
- 涉及的类包括KubernetesTagHostHandler、K8sServiceRoleInfo、KubernetesAbstractHandlerStrategy等，确保所有相关逻辑均能正确处理集群ID。

这些改动显著增强了数据模型的健壮性与一致性，确保了整体设计的一致性。

+++ datasophon-kubernetes/src/main/java/com/datasophon/kubernetes/actor/handler/KubernetesTagHostHandler.java
@@ -67,7 +67,7 @@ public class KubernetesTagHostHandler {
      * @param commandType 标签操作类型
      * @return 执行结果
      */
-    public ExecResult operateTag(Integer clusterId, String hostName, String kubeConfig,
+    public ExecResult operateTag(Long clusterId, String hostName, String kubeConfig,
             CommandType commandType) {
         ExecResult execResult = new ExecResult();
 
@@ -191,7 +191,7 @@ public class KubernetesTagHostHandler {
     /**
      * 获取目标主机列表
      */
-    private JSONArray getTargetHosts(Integer clusterId) {
+    private JSONArray getTargetHosts(Long clusterId) {
         Object obj = CacheUtils.get(clusterId + UNDERLINE + SERVICE_ROLE_HOST_MAPPING);
         if (obj == null) {
             return null;
