+++ datasophon-kubernetes/src/main/java/com/datasophon/kubernetes/strategy/KubernetesZkServerHandlerStrategy.java
@@ -42,7 +42,7 @@ public class KubernetesZkServerHandlerStrategy extends KubernetesAbstractHandler
     }
 
     @Override
-    public void getConfig(Integer clusterId, String namespace, List<ServiceConfig> list) {
+    public void getConfig(Long clusterId, String namespace, List<ServiceConfig> list) {
         if (list == null || list.isEmpty()) {
             return;
         }
